feat(ManagerContainer): Create manager container
