+++ app/src/components/forms/index.ts
@@ -1,0 +1,3 @@
+import PoolSelectionInput from "./PoolSelectionInput";
+
+export { PoolSelectionInput };
