+++ app/src/containers/index.ts
@@ -1,7 +1,9 @@
 import InvestorContainer from './InvestorContainer';
+import ManagerContainer from './ManagerContainer';
 import ProviderContainer from "./ProviderContainer";
 
 export {
   InvestorContainer,
+  ManagerContainer,
   ProviderContainer,
 }
